+++ Lib/fontbakery/fonts_profile.py
@@ -3,7 +3,7 @@ Font Bakery CheckRunner is the driver of a font bakery suite of checks.
 
 
 """
-from fontbakery.checkrunner import Profile
+from fontbakery.profile import Profile
 
 from fontbakery.callable import FontBakeryExpectedValue as ExpectedValue
