+++ Lib/fontbakery/profiles/ufo_sources.py
@@ -2,7 +2,9 @@ import os
 
 from fontbakery.callable import check, condition
 from fontbakery.callable import FontBakeryExpectedValue as ExpectedValue
-from fontbakery.checkrunner import ERROR, FAIL, PASS, WARN, Section, Profile
+from fontbakery.checkrunner import ERROR, FAIL, PASS, WARN
+from fontbakery.section import Section
+from fontbakery.profile import Profile
 
 BLACKLISTS = {}
